+++ packages/lumx-vue/src/components/text-field/index.ts
@@ -1,2 +1,3 @@
 export { default as RawInputText, type RawInputTextProps } from './RawInputText';
 export { default as RawInputTextarea, type RawInputTextareaProps } from './RawInputTextarea';
+export { default as TextField, type TextFieldProps } from './TextField';