+++ packages/site-demo/content/product/components/text-field/vue/icon.vue
@@ -1,0 +1,13 @@
+<template>
+    <TextField :value="value" :icon="mdiMagnify" :theme="theme" @change="value = $event" />
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue';
+import { mdiMagnify } from '@lumx/icons';
+import { TextField, type Theme } from '@lumx/vue';
+
+defineProps<{ theme?: Theme }>();
+
+const value = ref('');
+</script>