+++ packages/site-demo/content/product/components/text-field/vue/number.vue
@@ -1,0 +1,12 @@
+<template>
+    <TextField label="Number" :value="value" type="number" :theme="theme" @change="value = $event" />
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue';
+import { TextField, type Theme } from '@lumx/vue';
+
+defineProps<{ theme?: Theme }>();
+
+const value = ref('');
+</script>