+++ packages/site-demo/content/product/components/text-field/vue/placeholder.vue
@@ -1,0 +1,18 @@
+<template>
+    <TextField
+        label="Text field label"
+        :value="value"
+        placeholder="Placeholder text"
+        :theme="theme"
+        @change="value = $event"
+    />
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue';
+import { TextField, type Theme } from '@lumx/vue';
+
+defineProps<{ theme?: Theme }>();
+
+const value = ref('');
+</script>